# api/headgate.openapi.yaml
openapi: 3.1.0
info:
  title: headgate control API
  version: 0.1.0
  description: |
    The web UI is one client of this control API and gets no
    privileged access — asynqmon reads Redis directly, which is why its compatibility
    note is three minor versions stale.

    Both the Go and Rust implementations serve this spec, and the conformance suite
    asserts identical responses. Every list endpoint is bounded to prevent inspection
    from becoming an unbounded store operation: asynq's
    GetQueueInfo is O(number of groups) and has pinned Redis CPU for seconds in
    production. Monitoring caused the outage.

    DERIVED FROM THE ARCHITECTURE, NOT FROM A PREDECESSOR'S UI. The first draft of this
    spec was the asynq Inspector surface -- list by state, act on one job, pause a queue
    -- with new nouns bolted on. That inherited three problems: it omitted bulk
    operations, history, and enqueue (which asynq and apalis-board respectively DO have),
    and more importantly it had no endpoint for the question this system's own design
    creates. When dequeue is an admission decision, the operator's first question is not
    "what is in the queue" but "why is THIS job not running" -- see /jobs/{id}/admission.
    No predecessor needs that endpoint because no predecessor has a gate.
servers: [{url: /api/v1}]
paths:
...
  /jobs/{id}:
    get:
      summary: Job detail — attempts timeline, errors, and per-attempt logs.
      description: |
        Attempt history includes errors and logs saved at acknowledgement, not a live log stream.
        Each attempt's logs property remains an array of strings. Plain strings are info-level
        messages. Structured entries start with U+001E followed by "headgate-log-v1:" and a
        JSON object: level (debug/info/warn/error), at_ms (worker capture time), message,
        optional fields (scalar values), and optional truncated (boolean). Each encoded entry
        is bounded to 2048 UTF-8 bytes. Unknown versions or malformed entries should display
        as literal text. Log timestamps do not participate in store policy or lease decisions.
        Logs may contain application data and are not encrypted by payload encryption.
        `include_payload` applies to the job payload, not attempt logs.
      parameters:
        - {$ref: '#/components/parameters/Id'}
        - {name: include_payload, in: query, schema: {type: boolean, default: false}}
      responses: {'200': {description: OK}}
    delete: {parameters: [{$ref: '#/components/parameters/Id'}], responses: {'204': {description: Deleted}}}
...
components:
  parameters:
...
    Id:    {name: id,    in: path, required: true, schema: {type: string}}
...